requirements.txt
absl-py==2.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.7.9
charset-normalizer==3.4.2
chromadb==1.0.15
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
durationpy==0.10
exceptiongroup==1.3.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.5
fsspec==2025.5.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==3.10.0
kiwisolver==1.4.8
kubernetes==33.1.0
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
mmh3==5.1.0
mpmath==1.3.0
namex==0.1.0
narwhals==1.46.0
numpy==2.1.3
oauthlib==3.3.1
onnxruntime==1.22.1
opencv-python==4.12.0.88
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
opt_einsum==3.4.0
optree==0.16.0
orjson==3.10.18
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
posthog==5.4.0
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==310
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.46.1
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.21.2
toml==0.10.2
tomli==2.2.1
tornado==6.5.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.23.0
